eval: rewritten in cps
This will allow later on to implement 'pause' and 'continue'. Also, got
rid of the unevaluated representation of Logo code (the result type of
'expression' was 'unit -> atom'), since it is not necessary, since lists
are automatically quoted.  For example, in

> and [...] [...]

the code inside the brackets is not evaluated, so we do not have to
suspend evaluation of the arguments.  This simplifies everything a bit.